Extract cache key calculation from start parameter
