+++ subprojects/instant-execution/src/main/kotlin/org/gradle/instantexecution/InstantExecutionServices.kt
@@ -41,6 +41,7 @@ class InstantExecutionServices : AbstractPluginServiceRegistry() {
         registration.run {
             add(BuildTreeListenerManager::class.java)
             add(InstantExecutionStartParameter::class.java)
+            add(InstantExecutionCacheInvalidation::class.java)
             add(InstantExecutionReport::class.java)
             add(InstantExecutionProblems::class.java)
         }
